src/requirement.txt
mysql-connector-python
fastapi
uvicorn[standard]
